backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-generativeai==0.7.2
python-dotenv==1.0.0
Pillow==10.0.1
gunicorn==21.2.0
